requirements.txt
asgiref==3.8.1
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
cloudinary==1.41.0
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
Django==5.0.4
django-filter==23.5
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dulwich==0.21.7
fastjsonschema==2.20.0
ffmpeg-downloader==0.3.0
filelock==3.16.1
gTTS==2.5.1
gunicorn==23.0.0
idna==3.7
importlib_metadata==8.5.0
installer==0.7.0
jaraco.classes==3.4.0
jeepney==0.8.0
keyring==24.3.1
Markdown==3.6
more-itertools==10.5.0
msgpack==1.1.0
packaging==24.1
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.2
platformdirs==4.2.2
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
ptyprocess==0.7.0
pycparser==2.22
pydub==0.25.1
PyJWT==2.8.0
pyproject_hooks==1.2.0
pyttsx3==2.90
RapidFuzz==3.10.1
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
SecretStorage==3.3.3
shellingham==1.5.4
six==1.16.0
sqlparse==0.4.4
tomlkit==0.13.2
tqdm==4.66.4
trove-classifiers==2024.10.21.16
urllib3==2.2.2
virtualenv==20.27.1
whitenoise==6.8.2
zipp==3.20.2
